refactor: Rearrangement of files (#38)
* refactor: Rearrangement of files

Made changes in file arrangement.

* fix: minor changes

* fix: minor changes

* fix: minor changes

* fix: minor changes

* refactor: remove the extra home page, render home in index.js and actual app in frame.js

* refactor: rename components as per convention and use appropriate extensions

* fix: resolve build error with canvas

* fix: upgrade dependencies
