+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Frameboi",
+  "name": "frameboi",
   "private": true,
   "description": "Profile Picture Frame Generator for HackNITR 3.0",
   "version": "0.1.0",
@@ -33,11 +33,14 @@
     "@fortawesome/fontawesome-svg-core": "^1.2.36",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@fortawesome/react-fontawesome": "^0.1.15",
+    "babel-polyfill": "^6.26.0",
     "canvas": "^2.8.0",
     "gatsby": "^3.6.2",
     "gatsby-cli": "^3.12.0",
     "gatsby-plugin-robots-txt": "^1.6.8",
+    "gatsby-plugin-sharp": "^3.14.0",
     "gatsby-plugin-styled-components": "^4.10.0",
+    "gatsby-transformer-sharp": "^3.14.0",
     "konva": "^8.1.3",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
