+++ src/components/shared/Head.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
+
+// Libraries
 import styled from 'styled-components';
 import tw from 'twin.macro';
+
+// Components
 import home from '../../../config/home';
 
 const Heading = styled.div`
@@ -19,13 +23,11 @@ const Logo = styled.img`
     `}
 `;
 
-function Head() {
-  return (
-    <Heading>
-      <Logo src={home.logo.src} alt={home.logo.alt} />
-      <p>{home.head}</p>
-    </Heading>
-  );
-}
+const Head = () => (
+  <Heading>
+    <Logo src={home.logo.src} alt={home.logo.alt} />
+    <p>{home.head}</p>
+  </Heading>
+);
 
 export default Head;
