+++ src/pages/404.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Error = () => <div>Error 404: Page not found</div>;
+
+export default Error;
