+++ src/styles/global.css
@@ -148,7 +148,7 @@ table {
 /* Track */
 ::-webkit-scrollbar-track {
   background: #f5f5f5;
-  -webkit-box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
+  box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
   border: 1px solid #d5d5d5;
 }
 /* Handle */
